Delete README.md
The data conversion method proposed in this paper first obtains the latitude and longitude of each grid in the original irregular grid data, calculates the Euclidean distance between each grid point and the regular grid point, records the row and column numbers of the nearest point in terms of the Euclidean distance, and assigns the observed value corresponding to this nearest point to the regular grid. Finally, the mapping relationship between each regular grid node and the optimal irregular grid node is established. The advantage of this strategy is that for the same type of irregular grid data, once the regular grid is established, the spatial position of the nearest irregular grid node corresponding to each regular grid node is unchanged (the row and column numbers are unchanged), so the mapping relationship between them only needs to be calculated once. Then, the row and column numbers of each regular grid corresponding to the irregular grid can be directly obtained, and finally, the observed values can be extracted, which significantly improves the calculation efficiency.